# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cashop API
  version: 0.2.0
  summary: API REST de la plateforme de transferts Cashop (mobile et web).
  license:
    name: Propriétaire — tous droits réservés
    identifier: LicenseRef-Proprietary
  description: |
    Spécification **v0** (phase 1). Les montants sont des entiers en **unités mineures** ISO 4217
    (XAF : 0 décimale ; EUR, USD, GBP : 2 décimales).

    - Toutes les opérations financières exigent l'en-tête `Idempotency-Key`.
    - Chaque réponse renvoie `X-Correlation-Id` (généré s'il n'est pas fourni).
    - Aucun appel ne parle directement à un provider externe : l'orchestrateur Cashop s'en charge.
servers:
  - url: http://localhost:8080/api/v1
    description: Développement local
...
security:
  - bearerAuth: []
paths:
...
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [Auth]
      summary: Valider un OTP / code TOTP et obtenir un jeton
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code]
              properties:
                challenge_id: { type: string, format: uuid }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '200':
          description: Authentifié
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthToken' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /auth/mfa/totp:
    post:
      tags: [Auth]
      operationId: setupTotp
      summary: Démarrer l'activation de la MFA (application d'authentification)
      responses:
        '200':
          description: Secret à scanner (QR code de l'URI otpauth)
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string, description: Base32 }
                  otpauth_uri: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Auth]
      operationId: disableTotp
      summary: Désactiver la MFA (interdit au personnel)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TotpCode' }
      responses:
        '204': { description: MFA désactivée }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/mfa/totp/confirm:
    post:
      tags: [Auth]
      operationId: confirmTotp
      summary: Confirmer l'activation avec un premier code
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TotpCode' }
      responses:
        '204': { description: MFA activée }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/device/key:
    post:
      tags: [Auth]
      operationId: registerDeviceKey
      summary: Enregistrer la clé publique biométrique de l'appareil courant
      description: Clé EC P-256 (recommandée) ou RSA ≥ 2048, au format PEM, générée dans le Secure Enclave / Keystore.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_key]
              properties:
                public_key: { type: string }
      responses:
        '204': { description: Clé enregistrée }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/device/challenge:
    post:
      tags: [Auth]
      operationId: deviceChallenge
      summary: Obtenir un challenge à signer avec la clé biométrique (usage unique, 2 min)
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge_id: { type: string, format: uuid }
                  nonce: { type: string, description: Base64 — à signer en ECDSA/RSA SHA-256 }
                  expires_at: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Jeton Sanctum lié à un appareil (portée `customer` ou `staff`, durée limitée).
        Les apps web et admin ne l'exposent jamais au navigateur : leur BFF le conserve côté serveur.
...
  responses:
    ValidationError:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Interdit (KYC insuffisant, risque, permission)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Conflit (clé d'idempotence, statut incompatible, quote expirée)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: Trop de requêtes
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [VALIDATION_ERROR, UNAUTHENTICATED, INVALID_CREDENTIALS, ACCOUNT_LOCKED, INVALID_OTP, PIN_INVALID, PIN_LOCKED, MFA_REQUIRED, STAFF_ONLY, TOO_MANY_REQUESTS, KYC_REQUIRED, RISK_REVIEW, LIMIT_EXCEEDED, INSUFFICIENT_FUNDS, QUOTE_EXPIRED, IDEMPOTENCY_CONFLICT, REQUEST_IN_PROGRESS, PROVIDER_UNAVAILABLE]
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        correlation_id: { type: string }
...
    CountryCode:
      type: string
      pattern: '^[A-Z]{2}$'
...
    TotpCode:
      type: object
      required: [code]
      properties:
        code: { type: string, pattern: '^[0-9]{6}$' }
...
    AuthToken:
      type: object
      properties:
        token: { type: string }
        token_type: { type: string, const: Bearer }
        expires_at: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/Profile' }
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        phone: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        country: { $ref: '#/components/schemas/CountryCode' }
        email_verified: { type: boolean }
        phone_verified: { type: boolean }
        kyc_level: { type: string, examples: [LEVEL_1] }
        kyc_status: { $ref: '#/components/schemas/KycStatus' }
        mfa_enabled: { type: boolean }
        pin_set: { type: boolean }
...
    KycStatus:
      type: string
      enum: [PENDING, IN_REVIEW, VERIFIED, REJECTED, EXPIRED]
...